backend/requirements.txt
## Fichier généré par pip freeze le 2026-04-06 — version validée backend Overfitted.io
aiosqlite==0.22.1
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
certifi==2026.2.25
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet>=3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
Pillow==12.2.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-multipart==0.0.20
sniffio==1.3.1
sqladmin==0.24.0
SQLAlchemy==2.0.49
starlette==0.49.3
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
WTForms==3.1.2
celery==5.6.3
redis==7.4.0
python-dotenv==1.2.2
